Move the data parameter to string tests to it's own class and replace SimpleDateFormat with DateTimeFormatter.
