+++ core/src/main/java/org/frankframework/parameters/ParameterValue.java
@@ -18,6 +18,9 @@ package org.frankframework.parameters;
 import java.io.IOException;
 import java.text.DateFormat;
 import java.text.SimpleDateFormat;
+import java.time.ZoneId;
+import java.time.format.DateTimeFormatter;
+import java.time.format.DateTimeFormatterBuilder;
 import java.util.Collection;
 import java.util.Collections;
 import java.util.Date;
@@ -146,8 +149,8 @@ public class ParameterValue {
 		}
 
 		if (getDefinition() instanceof DateParameter dateParameter && value instanceof Date date) {
-			final DateFormat df = new SimpleDateFormat(dateParameter.getFormatString());
-			return df.format(date);
+			final DateTimeFormatter dtf = DateTimeFormatter.ofPattern(dateParameter.getFormatString()).withZone(ZoneId.systemDefault());
+			return dtf.format(date.toInstant());
 		}
 
 		return value.toString();
