+++ core/src/test/java/org/frankframework/parameters/DateParameterTest.java
@@ -425,50 +425,4 @@ public class DateParameterTest {
 		assertEquals(date, result);
 	}
 
-	private void testDateParameterToString(DateFormatType type, String pattern, String input, String expectedOutput) throws ParseException, ConfigurationException {
-		DateFormat df = new SimpleDateFormat(pattern);
-		Date date = df.parse(input);
-
-		DateParameter parameter = new DateParameter();
-		parameter.setName("InputMessage");
-		parameter.setFormatType(type);
-		parameter.configure();
-
-		ParameterValue pv = new ParameterValue(parameter, date);
-
-		assertInstanceOf(Date.class, pv.getValue());
-		assertEquals(pv.asStringValue(), expectedOutput);
-	}
-
-	@Test
-	public void testDateToString() throws ConfigurationException, ParseException {
-		testDateParameterToString(DateFormatType.DATE, TYPE_DATE_PATTERN, "2024-08-09 15:00:00", "2024-08-09");
-		testDateParameterToString(DateFormatType.DATE, TYPE_DATE_PATTERN, "2024-10-15", "2024-10-15");
-		testDateParameterToString(DateFormatType.DATE, TYPE_DATE_PATTERN, "1980-05-20 14:15:16.100", "1980-05-20");
-
-		assertThrows(ParseException.class, () -> testDateParameterToString(DateFormatType.DATE, TYPE_DATE_PATTERN, "20:00", ""));
-	}
-
-	@Test
-	public void testDateTimeToString() throws ConfigurationException, ParseException {
-		testDateParameterToString(DateFormatType.DATETIME, TYPE_DATETIME_PATTERN, "2024-08-09 18:15:25", "2024-08-09 18:15:25");
-		testDateParameterToString(DateFormatType.DATETIME, TYPE_DATETIME_PATTERN, "2024-08-09 00:00:00", "2024-08-09 00:00:00");
-
-		assertThrows(ParseException.class, () -> testDateParameterToString(DateFormatType.DATETIME, TYPE_DATETIME_PATTERN, "20:00", ""));
-	}
-
-	@Test
-	public void testTimeToString() throws ConfigurationException, ParseException {
-		testDateParameterToString(DateFormatType.TIME, TYPE_TIME_PATTERN, "19:25:45", "19:25:45");
-
-		assertThrows(ParseException.class, () -> testDateParameterToString(DateFormatType.TIME, TYPE_TIME_PATTERN, "2024-08-09 15:00", ""));
-	}
-
-	@Test
-	public void testTimestampToString() throws ConfigurationException, ParseException {
-		testDateParameterToString(DateFormatType.TIMESTAMP, TYPE_TIMESTAMP_PATTERN, "2024-08-09 18:15:25.500", "2024-08-09 18:15:25.500");
-
-		assertThrows(ParseException.class, () -> testDateParameterToString(DateFormatType.TIMESTAMP, TYPE_TIMESTAMP_PATTERN, "2024-08-09 15:00", ""));
-	}
-
 }
